requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-apscheduler==0.6.2
django-import-export==3.3.6
django-reversion==5.0.12
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
h11==0.14.0
idna==3.6
inflection==0.5.1
isodate==0.6.1
lxml==5.1.0
MarkupPy==1.14
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
PySocks==1.7.1
pyTelegramBotAPI==4.15.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tablib==3.5.0
telebot==0.0.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
webdriver-manager==4.0.1
whitenoise==6.6.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
zeep==4.2.1
